lint

+++ spatialmath/rotationMatrix.go
@@ -7,7 +7,7 @@ import (
 	"gonum.org/v1/gonum/num/quat"
 )
 
-// Rotation Matrix is a 3x3 matrix in row major order.
+// RotationMatrix is a 3x3 matrix in row major order.
 // m[3*r + c] is the element in the r'th row and c'th column.
 type RotationMatrix [9]float64
 
@@ -39,21 +39,21 @@ func (rm *RotationMatrix) AxisAngles() *R4AA {
 // reference: http://www.euclideanspace.com/maths/geometry/rotations/conversions/matrixToQuaternion/index.htm
 func (rm *RotationMatrix) Quaternion() quat.Number {
 	if tr := rm[0] + rm[4] + rm[8]; tr > 0 {
-		s := float64(0.5 / math.Sqrt(float64(tr+1.0)))
+		s := 0.5 / math.Sqrt(tr+1.0)
 		return quat.Number{0.25 / s, (rm[5] - rm[7]) * s, (rm[6] - rm[2]) * s, (rm[1] - rm[3]) * s}
 	}
 
 	if (rm[0] > rm[4]) && (rm[0] > rm[8]) {
-		s := float64(2.0 * math.Sqrt(float64(1.0+rm[0]-rm[4]-rm[8])))
+		s := 2.0 * math.Sqrt(1.0+rm[0]-rm[4]-rm[8])
 		return quat.Number{(rm[5] - rm[7]) / s, 0.25 * s, (rm[3] + rm[1]) / s, (rm[6] + rm[2]) / s}
 	}
 
 	if rm[4] > rm[8] {
-		s := float64(2.0 * math.Sqrt(float64(1.0+rm[4]-rm[0]-rm[8])))
+		s := 2.0 * math.Sqrt(1.0+rm[4]-rm[0]-rm[8])
 		return quat.Number{(rm[6] - rm[2]) / s, (rm[3] + rm[1]) / s, 0.25 * s, (rm[7] + rm[5]) / s}
 	}
 
-	s := float64(2.0 * math.Sqrt(float64(1.0+rm[8]-rm[0]-rm[4])))
+	s := 2.0 * math.Sqrt(1.0+rm[8]-rm[0]-rm[4])
 	return quat.Number{(rm[1] - rm[3]) / s, (rm[6] + rm[2]) / s, (rm[7] + rm[5]) / s, 0.25 * s}
 }
